LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/table64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:1"
FAIL #3 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:2"
FAIL #4 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:3"
FAIL #5 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:4"
FAIL #6 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:5"
FAIL #7 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:6"
FAIL #8 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:7"
FAIL #9 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:8"
FAIL #10 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:9"
FAIL #11 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:10"
FAIL #12 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:12"
FAIL #13 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table64.wast:13"
PASS #14 Test that WebAssembly compilation fails
PASS #15 Test that WebAssembly compilation fails
PASS #16 Reinitialize the default imports
FAIL #17 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #18 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #19 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #20 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:13:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #21 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #22 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:19:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #23 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #24 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:25:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #25 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #26 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:31:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #27 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #28 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:37:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #29 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #30 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:43:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #31 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #32 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:49:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #33 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #34 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #35 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:58:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #36 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 17: Memory64 is not enabled at {loc} expected true got false
FAIL #37 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:64:19
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
FAIL #38 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 35: Memory64 is not enabled at {loc} expected true got false
FAIL #39 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:70:19
global code@http://web-platform.test:8800/wasm/core/js/memory64/table64.wast.js:78:3 expected true got false
PASS #40 Test that WebAssembly compilation fails
PASS #41 Test that WebAssembly compilation fails
